fix lint script and improve build
