+++ src/App.test.tsx
@@ -3,7 +3,11 @@ import { render, screen } from "@testing-library/react";
 import { describe, it, expect } from "vitest";
 import Home from "./pages/Home";
 import { I18nextProvider } from "react-i18next";
-import i18n from "./i18n";
+import i18n, { initI18n } from "./i18n";
+
+beforeAll(async () => {
+  await initI18n();
+});
 
 describe("App", () => {
   it("renders the home page title", () => {
