+++ src/i18n.ts
@@ -1,19 +1,24 @@
 import i18n from "i18next";
 import { initReactI18next } from "react-i18next";
-import en from "./locales/en/translation.json";
-import fi from "./locales/fi/translation.json";
-import sv from "./locales/sv/translation.json";
 
-i18n.use(initReactI18next).init({
-  resources: {
-    en: { translation: en },
-    fi: { translation: fi },
-    sv: { translation: sv },
-  },
-  lng: "en",
-  fallbackLng: "en",
-  interpolation: { escapeValue: false },
-  react: { useSuspense: true },
-});
+export async function initI18n() {
+  const [en, fi, sv] = await Promise.all([
+    import("./locales/en/translation.json"),
+    import("./locales/fi/translation.json"),
+    import("./locales/sv/translation.json"),
+  ]);
+
+  await i18n.use(initReactI18next).init({
+    resources: {
+      en: { translation: en.default },
+      fi: { translation: fi.default },
+      sv: { translation: sv.default },
+    },
+    lng: "en",
+    fallbackLng: "en",
+    interpolation: { escapeValue: false },
+    react: { useSuspense: true },
+  });
+}
 
 export default i18n;
