+++ src/main.tsx
@@ -5,7 +5,7 @@ import App from "./App";
 import reportWebVitals from "./reportWebVitals";
 import { ThemeProvider } from "@dt/ThemeProvider";
 import { HelmetProvider } from "react-helmet-async";
-import "./i18n";
+import { initI18n } from "./i18n";
 
 const redirectPath = sessionStorage.getItem("redirectPath");
 if (redirectPath) {
@@ -31,18 +31,20 @@ if (gaId) {
   document.head.appendChild(inlineScript);
 }
 
-const root = ReactDOM.createRoot(
-  document.getElementById("root") as HTMLElement,
-);
-root.render(
-  <React.StrictMode>
-    <HelmetProvider>
-      <ThemeProvider>
-        <App />
-      </ThemeProvider>
-    </HelmetProvider>
-  </React.StrictMode>,
-);
+initI18n().then(() => {
+  const root = ReactDOM.createRoot(
+    document.getElementById("root") as HTMLElement,
+  );
+  root.render(
+    <React.StrictMode>
+      <HelmetProvider>
+        <ThemeProvider>
+          <App />
+        </ThemeProvider>
+      </HelmetProvider>
+    </React.StrictMode>,
+  );
+});
 
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
